Refactored CartItemsList.ts
