+++ src/services/useShopItemsList.ts
